[refac] MapViewController에서 PlaceListViewController 로직 extension으로 파일 분리
