+++ Project17-C-Map/InteractiveClusteringMap/Map/ViewController/MapViewController.swift
@@ -11,18 +11,18 @@ import NMapsMap
 
 final class MapViewController: UIViewController {
     
-    @IBOutlet weak var interactiveMapView: InteractiveMapView!
-    @IBOutlet private weak var placeListButton: UIButton!
+    @IBOutlet private(set) weak var interactiveMapView: InteractiveMapView!
+    @IBOutlet private(set) weak var placeListButton: UIButton!
     @IBOutlet private weak var editModeLabel: UILabel!
     @IBOutlet private weak var playButton: UIButton!
     
     private let locationManager = CLLocationManager()
     private var mapController: MapController?
     private var dataManager: DataManagable?
-    private var presentedMarkers: [NMFMarker] = []
+    private(set) var presentedMarkers: [NMFMarker] = []
     private var pickedMarker: LeafNodeMarker?
     private let leafNodeMarkerInfoWindow = LeafNodeMarkerInfoWindow()
-    private var placeListViewController: PlaceListViewController?
+    private(set) var placeListViewController: PlaceListViewController?
     private var isTouchedRemove: Bool = false
     
     init?(coder: NSCoder, dataManager: DataManagable) {
@@ -39,8 +39,8 @@ final class MapViewController: UIViewController {
         locationManager.requestWhenInUseAuthorization()
         dependencyInject()
         configureMap()
-        configurePlaceListViewController()
         bindingHandler()
+        configurePlaceListViewController()
     }
     
     override func viewDidAppear(_ animated: Bool) {
@@ -49,6 +49,16 @@ final class MapViewController: UIViewController {
         if !UserDefaultsManager.shared.isLaunched {
             presentOnboarding()
         }
+    }
+    
+    private func configurePlaceListViewController() {
+        guard let dataManager = dataManager,
+              let placeListViewController = placeListViewController(dataManager: dataManager) else { return }
+        
+        self.placeListViewController = placeListViewController
+        placeListViewController.didMove(toParent: self)
+        addChild(placeListViewController)
+        view.addSubview(placeListViewController.view)
     }
     
     private func presentOnboarding() {
@@ -75,10 +85,12 @@ final class MapViewController: UIViewController {
     private func configureMap() {
         interactiveMapView.mapView.addCameraDelegate(delegate: self)
         interactiveMapView?.mapView.touchDelegate = self
-        interactiveMapView.mapView.moveCamera(NMFCameraUpdate(scrollTo: NMGLatLng(lat: 37.56825785, lng: 126.9930027), zoomTo: 15))
         interactiveMapView.configureGesture()
         editModeLabel.isHidden = true
-        
+        configurePlayButton()
+    }
+    
+    private func configurePlayButton() {
         playButton.setImage(UIImage(systemName: "play.circle.fill")?.withTintColor(.blue), for: .normal)
         playButton.contentVerticalAlignment = .fill
         playButton.contentHorizontalAlignment = .fill
@@ -107,7 +119,6 @@ final class MapViewController: UIViewController {
     }
     
     private func addLeafNodeMarker(latlng: NMGLatLng) {
-        
         if interactiveMapView.zoomLevel > 15 {
         
         let alert = MapAlertControllerFactory.createAddAlertController { [weak self] text in
@@ -148,7 +159,7 @@ final class MapViewController: UIViewController {
         }
     }
     
-    func removeLeafNode(point: CGPoint) {
+    private func removeLeafNode(point: CGPoint) {
         for (index, marker) in presentedMarkers.enumerated() {
             guard let leafNodeMarker = marker as? LeafNodeMarker,
                   leafNodeMarker.containsMarker(at: point) else { continue }
@@ -196,7 +207,6 @@ final class MapViewController: UIViewController {
                 self.animate(marker: clusteringMarker)
             }
         }
-        
         updatePlaceListViewController()
     }
     
@@ -229,7 +239,7 @@ final class MapViewController: UIViewController {
         marker.animate(position: position)
     }
     
-    func setMarkersHandler(marker: ClusteringMarker) {
+    private func setMarkersHandler(marker: ClusteringMarker) {
         marker.touchHandler = { [weak self] _ in
             guard let self = self,
                   self.interactiveMapView.mode != .edit else { return true }
@@ -263,6 +273,10 @@ final class MapViewController: UIViewController {
             return true
         }
     }
+
+    @IBAction private func touchedPlayButton(_ sender: UIButton) {
+        interactiveMapView.playCameraAnimation()
+    }
     
 }
 
@@ -281,142 +295,6 @@ extension MapViewController: NMFMapViewTouchDelegate {
         }
         
         pickedMarker?.resizeMarkerSize()
-    }
-    
-}
-
-extension MapViewController {
-    
-    private func markersToCoordinates(_ markers: [NMFMarker]) -> [Coordinate] {
-        let clusters: [[Coordinate]] = markers.compactMap {
-            if let marker = $0 as? ClusteringMarker {
-                return marker.cluster.coordinates
-            }
-            
-            if let marker = $0 as? LeafNodeMarker {
-                return [marker.coordinate]
-            }
-            
-            return nil
-        }
-        return clusters.flatMap { $0 }
-    }
-    
-    @IBAction private func placeListButtonTouched(_ sender: UIButton) {
-        placeListButtonDisappearAnimation()
-        placeListViewController?.appearAnimation()
-        
-        let coordinates: [Coordinate] = markersToCoordinates(presentedMarkers)
-        let cluster = Cluster(coordinates: coordinates, boundingBox: .korea)
-        placeListViewController?.requestPlaces(cluster: cluster)
-    }
-    
-    @IBAction func touchedPlayButton(_ sender: UIButton) {
-        interactiveMapView.playCameraAnimation()
-    }
-    
-    private func updatePlaceListViewController() {
-        let coordinates: [Coordinate] = markersToCoordinates(presentedMarkers)
-        
-        if coordinates.count < 100 {
-            if placeListButton.isHidden && placeListViewController!.isShow == false {
-                placeListButtonAppear()
-            } else if placeListButton.isHidden {
-                let cluster = Cluster(coordinates: coordinates, boundingBox: .korea)
-                placeListViewController?.requestPlaces(cluster: cluster)
-            }
-        } else {
-            placeListButtonDisappear()
-            placeListViewController?.disappear()
-        }
-    }
-    
-    func placeListViewControllerDisappear() {
-        if placeListViewController?.isShow ?? false {
-            placeListButtonAppear()
-            placeListViewController?.disappear()
-            placeListViewController?.cancelAllOperation()
-        }
-    }
-    
-    private func placeListButtonAppear() {
-        placeListButton.isHidden = false
-        placeListButton.layer.isHidden = false
-        placeListButton.layer.opacity = 1
-    }
-    
-    private func placeListButtonDisappear() {
-        placeListButton.isHidden = true
-        placeListButton.layer.isHidden = true
-        placeListButton.layer.opacity = 0
-    }
-
-    private func placeListButtonDisappearAnimation() {
-        guard !placeListButton.isHidden else { return }
-        CATransaction.begin()
-        placeListButton.layer.opacity = 0
-        CATransaction.setCompletionBlock {
-            self.placeListButton.layer.isHidden = true
-            self.placeListButton.isHidden = true
-        }
-        let animation = AnimationController.floatingButtonAnimation(option: .disapper)
-        placeListButton.layer.add(animation, forKey: "floatingButtonDisappearAnimation")
-        CATransaction.commit()
-    }
-    
-    private func placeListButtonAppearAnimation() {
-        guard placeListButton.isHidden else { return }
-        CATransaction.begin()
-        placeListButton.layer.opacity = 1
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.2) {
-            self.placeListButton.layer.isHidden = false
-            self.placeListButton.isHidden = false
-        }
-        let animation = AnimationController.floatingButtonAnimation(option: .appear)
-        placeListButton.layer.add(animation, forKey: "floatingButtonAppearAnimation")
-        CATransaction.commit()
-    }
-    
-}
-
-private extension MapViewController {
-    
-    func configurePlaceListViewController() {
-        guard let dataManager = dataManager else { return }
-        
-        let poiService = POIService(dataManager: dataManager)
-        let geo = GeocodingNetwork(store: Store.http.dataProvider)
-        let img = ImageProvider(localStore: Store.local.dataProvider, httpStore: Store.http.dataProvider)
-        let service = PlaceInfoService(imageProvider: img, geocodingNetwork: geo)
-        
-        let storyboard = UIStoryboard(name: "Main", bundle: nil)
-        
-        placeListViewController = storyboard.instantiateViewController(
-            identifier: "PlaceListViewController",
-            creator: { coder in
-                return PlaceListViewController(coder: coder, poiService: poiService, placeInfoService: service)
-            })
-        guard let placeListViewController = placeListViewController else { return }
-        
-        placeListViewController.cancelButtonTouchedHandler = placeListButtonAppearAnimation
-        placeListViewController.delegate = self
-        placeListViewController.didMove(toParent: self)
-        addChild(placeListViewController)
-        view.addSubview(placeListViewController.view)
-    }
-    
-}
-
-extension MapViewController: PlaceListViewControllerDelegate {
-    
-    func selectedPlace(_ placeListViewController: PlaceListViewController, place: Place) {
-        let lat = place.coordinate.y
-        let lng = place.coordinate.x
-        let zoom: Double = interactiveMapView.mapView.maxZoomLevel
-        let cameraUpdate = NMFCameraUpdate(scrollTo: NMGLatLng(lat: lat, lng: lng), zoomTo: zoom)
-        cameraUpdate.animation = .fly
-        interactiveMapView.mapView.moveCamera(cameraUpdate)
-        placeListViewController.minimumHeightMode()
     }
     
 }
